IntHistogram Object and Class
