+++ raster/src/main/scala/geotrellis/raster/histogram/ArrayHistogram.scala
@@ -43,7 +43,7 @@ object ArrayHistogram {
   * Data object representing a histogram that uses an array for internal storage.
   */
 class ArrayHistogram(val counts: Array[Int], var total: Int)
-    extends MutableHistogramInt {
+    extends MutableIntHistogram {
   def size = counts.length
 
   def getTotalCount = total
